petavatar-agent/pyproject.toml
[project]
name = "petavatar-agent"
version = "0.1.0"
description = "AI agent for transforming pet photos into professional human avatars"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "bedrock-agentcore>=0.1.0",
    "boto3>=1.35.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
]
